requirements.txt
Pillow==9.5.0
pip==22.3.1
setuptools==65.5.1
wheel==0.38.4
